aqzae

+++ ejbModule/com/supinfo/interfaces/ICoursesServices.java
@@ -12,9 +12,9 @@ import com.supinfo.entity.User;
 @Remote
 public interface ICoursesServices {
 	public String authenticate(@WebParam(name = "login") String login, @WebParam(name = "mdp") String mdp );
-	public User getUser(Long idUser);
+	public User getUser(int idUser);
 	public List<Cours> getCours();
-	public Cours getCoursById(Long idCours);
-	public Cours takeCours(Long idCours,Long idUser);
-	public List<Question> passQuizz(Long idCours,Long idUser);
+	public Cours getCoursById(int idCours);
+	public Cours takeCours(int idCours,int idUser);
+	public List<Question> passQuizz(int idCours,int idUser);
 }
